% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEND.R
\name{alarm_estim}
\alias{alarm_estim}
\title{Estimate Tn}
\usage{
alarm_estim(newx, X, eps)
}
\arguments{
\item{newx}{a vector for the new observation.}

\item{X}{data matrix contains n observations and p columns.}

\item{eps}{a positive scalar which defines the neighborhood of each data point.}
}
\value{
The ratio of Tn to eps
}
\description{
Estimate Tn
}
