% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEND.R
\name{cv.smooth}
\alias{cv.smooth}
\title{Perform cross-validation smoothing to estimate the threshold}
\usage{
cv.smooth(X, lo, hi, len, alpha, eps.data)
}
\arguments{
\item{X}{data matrix contains n observations and p columns.}

\item{lo}{the smallest value for eps in the grid.}

\item{hi}{the largest value for eps in the grid.}

\item{len}{the number of points considered in the grid.}

\item{alpha}{a vector of quantiles for eps.}

\item{eps.data}{a positive scalar which defines the neighborhood of each data point.}
}
\value{
the list of chosen thresholds and the empirical probabilities.
}
\description{
Perform cross-validation smoothing to estimate the threshold
}
