% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEND.R
\name{eps_estim}
\alias{eps_estim}
\title{Estimate eps from the data}
\usage{
eps_estim(X)
}
\arguments{
\item{X}{data matrix contains n observations and p columns.}
}
\value{
The smallest radius eps such that the dataset is connected in the topological space.
}
\description{
Estimate eps from the data
}
