% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEND.R
\name{sm.boot}
\alias{sm.boot}
\title{Perform smoothed boostrap method to estimate the threshold}
\usage{
sm.boot(X, eps, N.boot = 10000)
}
\arguments{
\item{X}{data matrix contains n observations and p columns.}

\item{eps}{a positive scalar which defines the neighborhood of each data point.}

\item{N.boot}{the number of bootstrap samples.}
}
\description{
Perform smoothed boostrap method to estimate the threshold
}
